Refactor, move sql functions to sql_wrangle module
